Moving NpmDependenciesLinkProvider to its own file
